+++ app/api/e2ee.ts
@@ -6,7 +6,7 @@ import EncryptedKeyPair from "./models/encrypted_keypair.ts";
 import { saveObject } from "./services/unified_store.ts";
 import Account from "./models/account.ts";
 import authRequired from "./utils/auth.ts";
-import { getEnv } from "./utils/env_store.ts";
+import { getEnv } from "../shared/config.ts";
 import { rateLimit } from "./utils/rate_limit.ts";
 import {
   type ActivityPubActor,
