+++ app/api/login.ts
@@ -3,7 +3,7 @@ import { setCookie } from "hono/cookie";
 import { compare } from "bcrypt"; // bcrypt で検証
 import { z } from "zod";
 import { zValidator } from "@hono/zod-validator";
-import { getEnv } from "./utils/env_store.ts";
+import { getEnv } from "../shared/config.ts";
 import Session from "./models/session.ts";
 
 const app = new Hono();
