+++ app/api/utils/auth.ts
@@ -1,6 +1,6 @@
 import { MiddlewareHandler } from "hono";
 import Session from "../models/session.ts";
-import { getEnv } from "./env_store.ts";
+import { getEnv } from "../../shared/config.ts";
 import { createAuthMiddleware } from "../../shared/auth.ts";
 
 const authRequired: MiddlewareHandler = createAuthMiddleware({
