+++ app/takos_host/mailer.ts
@@ -1,6 +1,6 @@
-import { load } from "@std/dotenv";
+import { loadConfig } from "../../shared/config.ts";
 import * as nodemailer from "nodemailer";
-const env = await load();
+const env = await loadConfig();
 const smtp_host = env["SMTP_HOST"];
 const smtp_port = env["SMTP_PORT"];
 const smtp_auth_user = env["SMTP_USER"];
@@ -52,12 +52,12 @@ export const sendEmail = async (
   }
 };
 
-
 export const sendVerifyMail = async (
   to: string,
   code: string,
 ): Promise<boolean> => {
   const subject = "Takos Host メールアドレス確認";
-  const body = `以下のコードを入力してメールアドレスを確認してください。\n\n${code}`;
+  const body =
+    `以下のコードを入力してメールアドレスを確認してください。\n\n${code}`;
   return await sendEmail(to, subject, body);
 };
