+++ scripts/create_session_indexes.ts
@@ -1,9 +1,9 @@
-import { load } from "jsr:@std/dotenv";
+import { loadConfig } from "../shared/config.ts";
 import { connectDatabase } from "../shared/db.ts";
 import Session from "../app/api/models/session.ts";
 import HostSession from "../app/takos_host/models/session.ts";
 
-const env = await load();
+const env = await loadConfig();
 await connectDatabase(env);
 
 await Session.collection.createIndex({ expiresAt: 1 }, {
